Feature/replication

See merge request cpd/2122/t06/g01!15

+++ assign2/src/src/store/requests/GetRequestHandler.java
@@ -1,6 +1,7 @@
 package store.requests;
 
 import requests.GetRequest;
+import requests.SeekRequest;
 import store.state.NodeState;
 
 import java.io.IOException;
@@ -11,6 +12,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
 
 public class GetRequestHandler extends RequestHandler {
     public GetRequestHandler(NodeState nodeState) {
@@ -21,19 +23,32 @@ public class GetRequestHandler extends RequestHandler {
     void execute(String[] headers, OutputStream responseStream, InputStream clientData) throws IOException {
         GetRequest request = GetRequest.fromNetworkStream(headers);
         Path filePath = Paths.get(String.format("store-persistent-storage/%s/%s", getNodeState().getNodeId(), request.getKey()));
-        String neighbourId = getNeighbourhoodAlgorithms().findRequestDest(request.getKey());
-        if (neighbourId.equals(getNodeState().getNodeId())) {
-            if (!Files.exists(filePath)) {
-                responseStream.write("ERROR: Key not found\n".getBytes(StandardCharsets.UTF_8));
-            } else {
-                Files.copy(filePath, responseStream);
-            }
-        } else {
-            Socket neighbourSocket = new Socket(neighbourId, 3000);
-            request.send(neighbourSocket.getOutputStream());
-            neighbourSocket.getInputStream().transferTo(responseStream);
-            neighbourSocket.close();
+        List<String> allDest = getNeighbourhoodAlgorithms().findReplicationNodes(request.getKey());
+        if (allDest.contains(getNodeState().getNodeId()) && Files.exists(filePath)) {
+            Files.copy(filePath, responseStream);
+            return;
         }
-
+        // Removes the actual node because we already checked if the file existed
+        allDest.removeIf(nodeId -> nodeId.equals(getNodeState().getNodeId()));
+        SeekRequest seekRequest = new SeekRequest(request.getKey());
+        for (var nodeId: allDest) {
+            try {
+                Socket neighbourSocket = new Socket(nodeId, 3000);
+                OutputStream neighbourOutputStream = neighbourSocket.getOutputStream();
+                InputStream neighbourInputStream = neighbourSocket.getInputStream();
+                seekRequest.send(neighbourOutputStream);
+                String seekResponse = new String(neighbourInputStream.readAllBytes(), StandardCharsets.US_ASCII);
+                neighbourSocket.close();
+                if (seekResponse.equals(SeekRequest.FILE_FOUND_MESSAGE)) {
+                    neighbourSocket = new Socket(nodeId, 3000);
+                    request.send(neighbourSocket.getOutputStream());
+                    neighbourSocket.getInputStream().transferTo(responseStream);
+                    neighbourSocket.close();
+                    break;
+                }
+            } catch (IOException e) {
+                break;
+            }
+        }
     }
 }
